requirements.txt
python-binance==1.0.35
python-dotenv==1.0.1

